+++ lib/puppet_x/cups/server.rb
@@ -3,26 +3,20 @@ require 'open3'
 
 module PuppetX
   module Cups
+    # Namespace encapsulating helper functions
+    # to query the CUPS server
     module Server
-      # The IppResult class encapsules the use of the `ipptool` comand line utility.
-      # It provides helper methods and basic error handling.
-      class IppResult
-        def initialize(request, resource = '/')
-          @output, @error, process = query(resource, request)
-          raise('Unexpected response from CUPS server.') if process.exitstatus != 0
-        end
+      def self.query(request, resource = '/')
+        output = ipptool(resource, request)
+        output.split("\n")[1..-1]
+      rescue
+        raise("Failed to query resource 'ipp://localhost#{resource}' with request:\n#{request}")
+      end
 
-        def lines
-          @output.split("\n")[1..-1]
-        end
-
-        private
-
-        def query(resource, request)
-          Open3.capture3("ipptool -c ipp://localhost#{resource} /dev/stdin", stdin_data: request)
-        rescue
-          raise('Error using `ipptool` (CUPS 1.5 and later) command line tool.')
-        end
+      def self.ipptool(resource, request)
+        output, _, process = Open3.capture3("ipptool -c ipp://localhost#{resource} /dev/stdin", stdin_data: request)
+        raise if process.exitstatus != 0
+        output
       end
     end
   end